+++ tasty-query/shared/src/main/scala/tastyquery/TypeMaps.scala
@@ -1,13 +1,13 @@
-package tastyquery.ast
+package tastyquery
 
 import scala.collection.mutable
 
 import tastyquery.Contexts.*
-import tastyquery.ast.Flags.*
-import tastyquery.ast.Names.*
-import tastyquery.ast.Symbols.*
-import tastyquery.ast.Trees.*
-import tastyquery.ast.Types.*
+import tastyquery.Flags.*
+import tastyquery.Names.*
+import tastyquery.Symbols.*
+import tastyquery.Trees.*
+import tastyquery.Types.*
 
 object TypeMaps {
