+++ tasty-query/shared/src/main/scala/tastyquery/reader/classfiles/Descriptors.scala
@@ -1,15 +1,16 @@
 package tastyquery.reader.classfiles
 
-import tastyquery.Contexts.*
-import tastyquery.ast.Types
-import tastyquery.ast.Types.*
-import tastyquery.ast.Symbols.*
-import tastyquery.ast.Flags
-import tastyquery.reader.classfiles.ClassfileReader.ReadException
-import tastyquery.ast.Names.{nme, termName, typeName}
-import tastyquery.util.syntax.chaining.given
-
 import scala.annotation.switch
+
+import tastyquery.Contexts.*
+import tastyquery.Names.*
+import tastyquery.Types.*
+import tastyquery.Symbols.*
+import tastyquery.Flags
+
+import tastyquery.reader.classfiles.ClassfileReader.ReadException
+
+import tastyquery.util.syntax.chaining.given
 
 object Descriptors:
 
@@ -69,14 +70,14 @@ object Descriptors:
     def baseType: Option[Type] =
       if available >= 1 then
         (peek: @switch) match
-          case 'B' => commitSimple(1, Some(Types.ByteType))
-          case 'C' => commitSimple(1, Some(Types.CharType))
-          case 'D' => commitSimple(1, Some(Types.DoubleType))
-          case 'F' => commitSimple(1, Some(Types.FloatType))
-          case 'I' => commitSimple(1, Some(Types.IntType))
-          case 'J' => commitSimple(1, Some(Types.LongType))
-          case 'S' => commitSimple(1, Some(Types.ShortType))
-          case 'Z' => commitSimple(1, Some(Types.BooleanType))
+          case 'B' => commitSimple(1, Some(ByteType))
+          case 'C' => commitSimple(1, Some(CharType))
+          case 'D' => commitSimple(1, Some(DoubleType))
+          case 'F' => commitSimple(1, Some(FloatType))
+          case 'I' => commitSimple(1, Some(IntType))
+          case 'J' => commitSimple(1, Some(LongType))
+          case 'S' => commitSimple(1, Some(ShortType))
+          case 'Z' => commitSimple(1, Some(BooleanType))
           case _   => None
       else None
 
@@ -89,14 +90,14 @@ object Descriptors:
     def arrayType: Option[Type] =
       if consume('[') then
         val tpe = fieldDescriptor
-        Some(Types.ArrayTypeOf(tpe))
+        Some(ArrayTypeOf(tpe))
       else None
 
     def fieldDescriptor: Type =
       baseType.orElse(objectType).orElse(arrayType).getOrElse(abort)
 
     def returnDescriptor: Type =
-      if consume('V') then Types.UnitType
+      if consume('V') then UnitType
       else fieldDescriptor
 
     def methodDescriptor: Type =
